+++ packages/remix-netlify/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@remix-run/netlify`
 
+## 1.16.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.16.1`
+
 ## 1.16.0
 
 ### Patch Changes